Split class from tests
