--- td/challenge260/InstructionParser.h
@@ -1,5 +1,0 @@
-#pragma once
-class D
-{
-
-};
